requirements.txt
flask
openai
python-dotenv
cloudinary